moved component directory
